requirements-tox.txt

# Файл для обновления/сброса кеша в actions/cache
#
# actions/cache не имеет ручного механизма очистки кеша.
# Для сброса нужно менять ключ кеша, в качестве ключа кеширования
# используется хеш этого файла. Для сброса обновите дату.
#
# Обновлено: 23 мая 2022

cryptography # используется в libcloud для ssh-key fingerprint
pytest
pytest-cov
vcrpy
